+++ ethers-contract/src/contract.rs
@@ -1,6 +1,6 @@
 use crate::{
     base::{encode_function_data, AbiError, BaseContract},
-    call::ContractCall,
+    call::FunctionCall,
     event::{EthEvent, Event},
     EthLogDecode,
 };
@@ -9,12 +9,19 @@ use ethers_core::{
     types::{Address, Filter, Selector, ValueOrArray},
 };
 use ethers_providers::Middleware;
-use std::{marker::PhantomData, sync::Arc};
+use std::{borrow::Borrow, fmt::Debug, marker::PhantomData, sync::Arc};
 
 #[cfg(not(feature = "legacy"))]
 use ethers_core::types::Eip1559TransactionRequest;
 #[cfg(feature = "legacy")]
 use ethers_core::types::TransactionRequest;
+
+/// `Contract` is a [`ContractInstance`] object with an `Arc` middleware.
+/// This type alias exists to preserve backwards compatibility with
+/// less-abstract Contracts.
+///
+/// For full usage docs, see [`ContractInstance`].
+pub type Contract<M> = ContractInstance<std::sync::Arc<M>, M>;
 
 /// A Contract is an abstraction of an executable program on the Ethereum Blockchain.
 /// It has code (called byte code) as well as allocated long-term memory
@@ -69,7 +76,7 @@ use ethers_core::types::TransactionRequest;
 /// use ethers_contract::Contract;
 /// use ethers_providers::{Provider, Http};
 /// use ethers_signers::Wallet;
-/// use std::convert::TryFrom;
+/// use std::{convert::TryFrom, sync::Arc};
 ///
 /// # async fn foo() -> Result<(), Box<dyn std::error::Error>> {
 /// // this is a fake address used just for this example
@@ -82,7 +89,7 @@ use ethers_core::types::TransactionRequest;
 /// let client = Provider::<Http>::try_from("http://localhost:8545").unwrap();
 ///
 /// // create the contract object at the address
-/// let contract = Contract::new(address, abi, client);
+/// let contract = Contract::new(address, abi, Arc::new(client));
 ///
 /// // Calling constant methods is done by calling `call()` on the method builder.
 /// // (if the function takes no arguments, then you must use `()` as the argument)
@@ -116,13 +123,13 @@ use ethers_core::types::TransactionRequest;
 /// use ethers_contract::{Contract, EthEvent};
 /// use ethers_providers::{Provider, Http, Middleware};
 /// use ethers_signers::Wallet;
-/// use std::convert::TryFrom;
+/// use std::{convert::TryFrom, sync::Arc};
 /// use ethers_core::abi::{Detokenize, Token, InvalidOutputType};
 /// # // this is a fake address used just for this example
 /// # let address = "eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee".parse::<Address>()?;
 /// # let abi: Abi = serde_json::from_str(r#"[]"#)?;
 /// # let client = Provider::<Http>::try_from("http://localhost:8545").unwrap();
-/// # let contract = Contract::new(address, abi, client);
+/// # let contract = Contract::new(address, abi, Arc::new(client));
 ///
 /// #[derive(Clone, Debug, EthEvent)]
 /// struct ValueChanged {
@@ -145,18 +152,40 @@ use ethers_core::types::TransactionRequest;
 ///
 /// _Disclaimer: these above docs have been adapted from the corresponding [ethers.js page](https://docs.ethers.io/ethers.js/html/api-contract.html)_
 ///
+/// # Usage Note
+///
+/// `ContractInternal` accepts any client that implements `B: Borrow<M>` where
+/// `M :Middleware`. Previous `Contract` versions used only arcs, and relied
+/// heavily on [`Arc`]. Due to constraints on the [`FunctionCall`] type,
+/// calling contracts requires a `B: Borrow<M> + Clone`. This is fine for most
+/// middlware. However, when `B` is an owned middleware that is not Clone, we
+/// cannot issue contract calls. Some notable exceptions:
+///
+/// - `NonceManagerMiddleware`
+/// - `SignerMiddleware` (when using a non-Clone Signer)
+///
+/// When using non-Clone middlewares, instead of instantiating a contract that
+/// OWNS the middlware, pass the contract a REFERENCE to the middleware. This
+/// will fix the trait bounds issue (as `&M` is always `Clone`).
+///
+/// We expect to fix this fully in a future version
+///
 /// [`abigen`]: macro.abigen.html
 /// [`Abigen` builder]: struct.Abigen.html
-/// [`event`]: method@crate::Contract::event
-/// [`method`]: method@crate::Contract::method
+/// [`event`]: method@crate::ContractInstance::event
+/// [`method`]: method@crate::ContractInstance::method
 #[derive(Debug)]
-pub struct Contract<M> {
+pub struct ContractInstance<B, M> {
     address: Address,
     base_contract: BaseContract,
-    client: Arc<M>,
+    client: B,
+    _m: PhantomData<M>,
 }
 
-impl<M> std::ops::Deref for Contract<M> {
+impl<B, M> std::ops::Deref for ContractInstance<B, M>
+where
+    B: Borrow<M>,
+{
     type Target = BaseContract;
 
     fn deref(&self) -> &Self::Target {
@@ -164,18 +193,25 @@ impl<M> std::ops::Deref for Contract<M> {
     }
 }
 
-impl<M> Clone for Contract<M> {
+impl<B, M> Clone for ContractInstance<B, M>
+where
+    B: Clone + Borrow<M>,
+{
     fn clone(&self) -> Self {
-        Contract {
+        ContractInstance {
             base_contract: self.base_contract.clone(),
             client: self.client.clone(),
             address: self.address,
+            _m: self._m,
         }
     }
 }
 
-impl<M> Contract<M> {
-    /// Returns the contract's address.
+impl<B, M> ContractInstance<B, M>
+where
+    B: Borrow<M>,
+{
+    /// Returns the contract's address
     pub fn address(&self) -> Address {
         self.address
     }
@@ -186,17 +222,24 @@ impl<M> Contract<M> {
     }
 
     /// Returns a pointer to the contract's client.
-    pub fn client(&self) -> Arc<M> {
-        Arc::clone(&self.client)
+    pub fn client(&self) -> B
+    where
+        B: Clone,
+    {
+        self.client.clone()
     }
 
     /// Returns a reference to the contract's client.
     pub fn client_ref(&self) -> &M {
-        Arc::as_ref(&self.client)
+        self.client.borrow()
     }
 }
 
-impl<M: Middleware> Contract<M> {
+impl<B, M> ContractInstance<B, M>
+where
+    B: Borrow<M>,
+    M: Middleware,
+{
     /// Returns an [`Event`](crate::builders::Event) builder for the provided event.
     /// This function operates in a static context, then it does not require a `self`
     /// to reference to instantiate an [`Event`](crate::builders::Event) builder.
@@ -209,14 +252,14 @@ impl<M: Middleware> Contract<M> {
     }
 }
 
-impl<M: Middleware> Contract<M> {
+impl<B, M> ContractInstance<B, M>
+where
+    B: Borrow<M>,
+    M: Middleware,
+{
     /// Creates a new contract from the provided client, abi and address
-    pub fn new(
-        address: impl Into<Address>,
-        abi: impl Into<BaseContract>,
-        client: impl Into<Arc<M>>,
-    ) -> Self {
-        Self { base_contract: abi.into(), client: client.into(), address: address.into() }
+    pub fn new(address: impl Into<Address>, abi: impl Into<BaseContract>, client: B) -> Self {
+        Self { base_contract: abi.into(), client, address: address.into(), _m: PhantomData }
     }
 
     /// Returns an [`Event`](crate::builders::Event) builder for the provided event.
@@ -227,7 +270,7 @@ impl<M: Middleware> Contract<M> {
     /// Returns an [`Event`](crate::builders::Event) builder with the provided filter.
     pub fn event_with_filter<D: EthLogDecode>(&self, filter: Filter) -> Event<M, D> {
         Event {
-            provider: &self.client,
+            provider: self.client.borrow(),
             filter: filter.address(ValueOrArray::Value(self.address)),
             datatype: PhantomData,
         }
@@ -240,40 +283,49 @@ impl<M: Middleware> Contract<M> {
         Ok(self.event_with_filter(Filter::new().event(&event.abi_signature())))
     }
 
-    /// Returns a transaction builder for the provided function name. If there are
-    /// multiple functions with the same name due to overloading, consider using
-    /// the `method_hash` method instead, since this will use the first match.
-    pub fn method<T: Tokenize, D: Detokenize>(
-        &self,
-        name: &str,
-        args: T,
-    ) -> Result<ContractCall<M, D>, AbiError> {
-        // get the function
-        let function = self.base_contract.abi.function(name)?;
-        self.method_func(function, args)
+    /// Returns a new contract instance using the provided client
+    ///
+    /// Clones `self` internally
+    #[must_use]
+    pub fn connect<N>(&self, client: Arc<N>) -> ContractInstance<Arc<N>, N>
+    where
+        N: Middleware,
+    {
+        ContractInstance {
+            base_contract: self.base_contract.clone(),
+            client,
+            address: self.address,
+            _m: PhantomData,
+        }
     }
 
-    /// Returns a transaction builder for the selected function signature. This should be
-    /// preferred if there are overloaded functions in your smart contract
-    pub fn method_hash<T: Tokenize, D: Detokenize>(
-        &self,
-        signature: Selector,
-        args: T,
-    ) -> Result<ContractCall<M, D>, AbiError> {
-        let function = self
-            .base_contract
-            .methods
-            .get(&signature)
-            .map(|(name, index)| &self.base_contract.abi.functions[name][*index])
-            .ok_or_else(|| Error::InvalidName(hex::encode(signature)))?;
-        self.method_func(function, args)
+    /// Returns a new contract instance using the provided client
+    ///
+    /// Clones `self` internally
+    #[must_use]
+    pub fn connect_with<C, N>(&self, client: C) -> ContractInstance<C, N>
+    where
+        C: Borrow<N>,
+    {
+        ContractInstance {
+            base_contract: self.base_contract.clone(),
+            client,
+            address: self.address,
+            _m: PhantomData,
+        }
     }
+}
 
+impl<B, M> ContractInstance<B, M>
+where
+    B: Clone + Borrow<M>,
+    M: Middleware,
+{
     fn method_func<T: Tokenize, D: Detokenize>(
         &self,
         function: &Function,
         args: T,
-    ) -> Result<ContractCall<M, D>, AbiError> {
+    ) -> Result<FunctionCall<B, M, D>, AbiError> {
         let data = encode_function_data(function, args)?;
 
         #[cfg(feature = "legacy")]
@@ -291,13 +343,43 @@ impl<M: Middleware> Contract<M> {
 
         let tx = tx.into();
 
-        Ok(ContractCall {
+        Ok(FunctionCall {
             tx,
-            client: Arc::clone(&self.client), // cheap clone behind the Arc
+            client: self.client.clone(),
             block: None,
             function: function.to_owned(),
             datatype: PhantomData,
+            _m: self._m,
         })
+    }
+
+    /// Returns a transaction builder for the selected function signature. This should be
+    /// preferred if there are overloaded functions in your smart contract
+    pub fn method_hash<T: Tokenize, D: Detokenize>(
+        &self,
+        signature: Selector,
+        args: T,
+    ) -> Result<FunctionCall<B, M, D>, AbiError> {
+        let function = self
+            .base_contract
+            .methods
+            .get(&signature)
+            .map(|(name, index)| &self.base_contract.abi.functions[name][*index])
+            .ok_or_else(|| Error::InvalidName(hex::encode(signature)))?;
+        self.method_func(function, args)
+    }
+
+    /// Returns a transaction builder for the provided function name. If there are
+    /// multiple functions with the same name due to overloading, consider using
+    /// the `method_hash` method instead, since this will use the first match.
+    pub fn method<T: Tokenize, D: Detokenize>(
+        &self,
+        name: &str,
+        args: T,
+    ) -> Result<FunctionCall<B, M, D>, AbiError> {
+        // get the function
+        let function = self.base_contract.abi.function(name)?;
+        self.method_func(function, args)
     }
 
     /// Returns a new contract instance at `address`.
@@ -308,13 +390,5 @@ impl<M: Middleware> Contract<M> {
         let mut this = self.clone();
         this.address = address.into();
         this
-    }
-
-    /// Returns a new contract instance using the provided client
-    ///
-    /// Clones `self` internally
-    #[must_use]
-    pub fn connect<N>(&self, client: Arc<N>) -> Contract<N> {
-        Contract { base_contract: self.base_contract.clone(), client, address: self.address }
     }
 }