+++ CHANGELOG.md
@@ -297,6 +297,8 @@
 
 ### Unreleased
 
+- Make `Contract` objects generic over the borrow trait, to allow non-arc mware
+  [#2082](https://github.com/gakonst/ethers-rs/pull/2082)
 - Return pending transaction from `Multicall::send`
   [#2044](https://github.com/gakonst/ethers-rs/pull/2044)
 - Add abigen to default features